+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2023-08-27
+
+### Added
+
+- [enhancement](https://github.com/FrankC01/pysui-fastcrypto/issues/3) Sign arbirary message
+- [enhancement](https://github.com/FrankC01/pysui-fastcrypto/issues/2) Verify signature
+
+### Fixed
+
+### Changed
+
+### Removed
+
 ## [0.1.7] - 2023-08-23
 
 ### Added